# scripts/openapi.yaml
openapi: '3.1.0'
info:
  title: File Sync WebApp
  description: Web Frontend for File Sync Service
  contact:
    name: Daniel Boline
    email: ddboline@gmail.com
  license:
    name: ''
  version: '0.12.4'
paths:
  /sync/delete_cache_entry:
    delete:
      operationId: delete_cache_entry
      responses:
        '200':
          description: Delete Cache Entry
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
  /sync/index.html:
    get:
      operationId: sync_frontpage
      responses:
        '200':
          description: Main Page
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
  /sync/list_sync_cache:
    get:
      operationId: list_sync_cache
      responses:
        '200':
          description: List Sync Cache
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
  /sync/proc:
    post:
      operationId: process_cache_entry
      responses:
        '200':
          description: Process Entry
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
  /sync/proc_all:
    post:
      operationId: proc_all
      responses:
        '200':
          description: Process All
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
  /sync/remove:
    delete:
      operationId: remove
      responses:
        '200':
          description: Remove Sync Entry
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
...
  /sync/sync:
    post:
      operationId: sync_all
      responses:
        '200':
          description: Sync
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
  /sync/sync/{name}:
    post:
      operationId: sync_name
      responses:
        '200':
          description: Sync
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
  /sync/sync_calendar:
    post:
      operationId: sync_calendar
      responses:
        '200':
          description: Sync Calendar DB
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
  /sync/sync_garmin:
    post:
      operationId: sync_garmin
      responses:
        '200':
          description: Sync Garmin DB
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
  /sync/sync_movie:
    post:
      operationId: sync_movie
      responses:
        '200':
          description: Sync Movie DB
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
  /sync/sync_podcasts:
    post:
      operationId: sync_podcasts
      responses:
        '200':
          description: Sync Podcasts
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
  /sync/sync_security:
    post:
      operationId: sync_security
      responses:
        '200':
          description: Sync Security Logs
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
  /sync/sync_weather:
    post:
      operationId: sync_weather
      responses:
        '200':
          description: Sync Weather Data
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
...
components:
  schemas:
...
    str:
      type: string